advance: cosine loss in fp32 for stability (FID 315.19 vs 317.0)

+++ train.py
@@ -1065,7 +1065,7 @@ while True:
                 x_t, velocity = flow_matching.forward_sample(x, t)
                 pred = model(x_t, t, class_labels=y)
                 huber = F.smooth_l1_loss(pred, velocity)
-                cos = 1.0 - F.cosine_similarity(pred.flatten(1), velocity.flatten(1), dim=1).mean()
+                cos = 1.0 - F.cosine_similarity(pred.flatten(1).float(), velocity.flatten(1).float(), dim=1).mean()
                 loss = huber + 0.5 * cos
             train_loss = huber.detach()
             loss = loss / grad_accum_steps
